refactor(workbench): extract sidebar session tree into WorkbenchSessionTree
Catalog paint (projects, orphans, multi-select) leaves AppWorkbench. Open/new-chat and UserMenu stay on the host.
